chore: extract reduction module to file
